requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
PyMySQL==1.1.1
cryptography==43.0.1
